* master: (706 commits)
  [DOCS] Updates security API examples (#10752)
  Update Logstash release for 6.8 branch to 6.8.7 (#11659)
  Update logstash_releases.json with 6.8.8 and 7.6.1 (#11658)
  Updated: JRuby to 9.2.11.0
  Fix: use l/w match-ing (which does not depend on frames)
  Fixed time format that introduced flaky tests depending on time
  Fix: broken --help due sinatra 2.x upgrade
  Update rack dependency on docgen project
  Fixes JDK13's javadoc build failure
Adaptations to internal collector to send data directly to monitoring
cluster Close 11573
  remove TODO from CONTRIBUTING notes
Fixes #11598 enabling the users to use completely commented config
files
  Remove deprecation notices
  Make capitalization consistent
  Fixes out-dated monitoring links
  Add shared attribute for cloud trial
  Update contributing guidelines to clarify changelog formatting
Changed PluginFactory to resolve id field with environment variables
docs: plugin ids variable expansion cannot use secret store
  Fix setting name for monitoring
  Updated releases.json after 7.6.0 release (#11595)
  ...

+++ ci/docker_run.sh
@@ -14,6 +14,10 @@ else
     IMAGE_NAME=$branch_specifier"-"$(date +%s%N)
 fi
 
+if [ "$OSS" == "true" ]; then
+  DOCKER_ENV_OPTS="${DOCKER_ENV_OPTS} --env OSS=true"
+fi
+
 echo "Running Docker CI build for '$IMAGE_NAME' "
 
 # Remove old docker cid just in case
@@ -23,14 +27,19 @@ docker build -t $IMAGE_NAME .
 exit_code=$?; [[ $exit_code != 0 ]] && exit $exit_code
 
 cleanup() {
-  cat docker_cid | xargs docker rm --force -v 
+  if [ -e docker_cid ]; then
+    cat docker_cid | xargs docker rm --force -v
+  fi
 }
 trap cleanup EXIT
 
 # Run the command, skip the first argument, which is the image name
-echo "Running tests in built docker image"
-docker run --sig-proxy=true --cidfile=docker_cid --rm $IMAGE_NAME ${@:2}
+docker run $DOCKER_ENV_OPTS --cidfile=docker_cid --sig-proxy=true --rm $IMAGE_NAME ${@:2}
 exit_code=$?
+
+# Remove the container cid since we ran cleanly, no need to force rm it if we got to this point
+rm docker_cid
+
 [[ $REMOVE_IMAGE == "true" ]] && docker rmi $IMAGE_NAME
 echo "exiting with code: '$exit_code'"
 exit $exit_code #preserve the exit code from the test run
